ocr-python/requirements.txt
flask
google-cloud-vision
opencv-python
pillow
numpy
